EntryListSearch component added
